Copy changes to project
